Get addon docs basically working.
