+++ packages/ember-principled-forms/addon/components/principled-form-field/template.hbs
@@ -3,8 +3,8 @@
   <input
     value={{value}}
     type={{type}}
-    onchange={{action 'change' value='target.value'}}
-    oninput={{action 'input' value='target.value'}}
+    onchange={{action 'handleChange' value='target.value'}}
+    oninput={{action 'handleInput' value='target.value'}}
     class={{inputClass}}
   >
 </label>
