+++ packages/ember-principled-forms/tests/dummy/app/templates/application.hbs
@@ -1,19 +1,3 @@
 {{outlet}}
 
 {{docs-keyboard-shortcuts}}
-
-{{principled-form-field
-    type=user.name.type
-    value=user.name.value
-    label='Name'
-    onChange=(action 'change' 'name')
-    onInput=(action 'input' 'name')
-}}
-
-{{principled-form-field
-    type=user.age.type
-    value=user.age.value
-    label='Age'
-    onChange=(action 'change' 'age')
-    onInput=(action 'input' 'age')
-}}
