+++ packages/ember-principled-forms/tests/dummy/app/templates/docs/components.hbs
@@ -1,0 +1,1 @@
+{{form-field}}
