--- packages/ember-principled-forms/tests/dummy/app/templates/docs/usage.hbs
@@ -1,1 +1,0 @@
-{{outlet}}
