+++ packages/ember-principled-forms/tests/dummy/app/templates/docs/usage.md
@@ -1,0 +1,3 @@
+```
+ember install ember-principled-forms
+```
